Minor: restructure cli tooling
